contract_analyzer/requirements.txt
openai
langchain
langchain-community
chromadb
pymupdf
python-dotenv

